Add IBCError component and update IBCSend component
